-) Corrected some integer - unsigned integer comparisons, which caused warnings in nightly-tests -) Problem when compiling with icc remains. I don't know why icc has problems with certain mkl-types e.g. sparse_matrix_t. Maybe the icc version is too old (2015) and can't handle types from mkl-2017?

+++ source/OLAS/multigrid/depgraph.cc
@@ -300,7 +300,7 @@ InsertStartArray( const UInt *const startarray,
         // if this object owns the start array
         if( ownStartIndex_ ) {
             // check, if the sizes match
-            if( size != NumNodes_ ) {
+            if( (Integer)size != NumNodes_ ) {
                 // cannot change number of nodes, if connections already exist
                 if( NumEdges_ )  return false;
                 delete [] ( StartIndex_ );  StartIndex_  = NULL;
